Add error handling and messaging

+++ index.js
@@ -10,15 +10,39 @@ const backupNamePrefix = "auto"; // param
 const deleteNamePrefix = "auto"; // param
 
 const managementToken = ""; // secret
+if (!managementToken) {
+    console.error(`MISSING MANAGEMENT TOKEN: See README`);
+    process.exit(1);
+}
+
 const spaceId = "nrocpvfo0sk3"; // param
+if (!spaceId) {
+    console.error(`MISSING SPACE ID: See README`);
+    process.exit(1);
+}
+
 const accessString = `--management-token ${managementToken} --space-id ${spaceId}`;
 
 const envs = getAndParseContentfulEnvs(accessString);
 
-const envToDelete = getEnvToDelete(envs, deleteNamePrefix, backupNameSuffix);
+if (!!deleteNamePrefix) {
+    const envToDelete = getEnvToDelete(
+        envs,
+        deleteNamePrefix,
+        backupNameSuffix
+    );
+    envToDelete && deleteEnv(envToDelete, accessString);
+} else {
+    console.warn(`SKIPPING DELETE: deleteNamePrefix not supplied`);
+}
 
-envToDelete && deleteEnv(envToDelete, accessString);
-
-const envToCreate = getBackupName(backupNamePrefix, new Date(), backupNameSuffix);
-
-createEnv(envToCreate, accessString);
+if (!!backupNamePrefix) {
+    const envToCreate = getBackupName(
+        backupNamePrefix,
+        new Date(),
+        backupNameSuffix
+    );
+    createEnv(envToCreate, accessString);
+} else {
+    console.warn(`SKIPPING CREATE: backupNamePrefix not supplied`);
+}
